DRY up sync_geth and ipfs_pull_hashes
